modify data reader to test curvature

+++ dataReader/dataReader.java
@@ -58,7 +58,7 @@ public class dataReader
                     float normal_z = (float)Double.parseDouble(coordinates[5]);
                     points.add(new point(x, y, z, normal_x, normal_y, normal_z));
                 } else if (coordinates.length == 4) {
-                    int color = (int)Double.parseDouble(coordinates[3]);
+                    float color = (float)Double.parseDouble(coordinates[3]);
                     points.add(new point(x, y, z, color));   
                 } else {
                     points.add(new point(x, y, z));
